federation troubleshooting
I have manually corrected as many federations as possible to refer to the correct iso3c code for the country to which the federation refers. This has dropped a few thousand players, and has combined England, Scotland, and Whales into the GBR code. This was done using Excel because I could not for the life of me get {rvest} to extract the hyperlinks from the FIDE table website which themselves contained the federation codes. Refers to #1